+++ src/ros_std.hpp
@@ -12,21 +12,21 @@
 #include <std_msgs/UInt32.h>
 #include <std_msgs/UInt64.h>
 
-long portRead(const ubx_port_t *p, char *x, const int len);
-long portRead(const ubx_port_t *p, float *x, const int len);
-long portRead(const ubx_port_t *p, double *x, const int len);
-long portRead(const ubx_port_t *p, int32_t *x, const int len);
-long portRead(const ubx_port_t *p, int64_t *x, const int len);
-long portRead(const ubx_port_t *p, uint32_t *x, const int len);
-long portRead(const ubx_port_t *p, uint64_t *x, const int len);
+long portRead(const ubx_port_t *p, char *x, const long len);
+long portRead(const ubx_port_t *p, float *x, const long len);
+long portRead(const ubx_port_t *p, double *x, const long len);
+long portRead(const ubx_port_t *p, int32_t *x, const long len);
+long portRead(const ubx_port_t *p, int64_t *x, const long len);
+long portRead(const ubx_port_t *p, uint32_t *x, const long len);
+long portRead(const ubx_port_t *p, uint64_t *x, const long len);
 
-long portWrite(const ubx_port_t *p, const char *x, const int len);
-long portWrite(const ubx_port_t *p, const float *x, const int len);
-long portWrite(const ubx_port_t *p, const double *x, const int len);
-long portWrite(const ubx_port_t *p, const int32_t *x, const int len);
-long portWrite(const ubx_port_t *p, const int64_t *x, const int len);
-long portWrite(const ubx_port_t *p, const uint32_t *x, const int len);
-long portWrite(const ubx_port_t *p, const uint64_t *x, const int len);
+long portWrite(const ubx_port_t *p, const char *x, const long len);
+long portWrite(const ubx_port_t *p, const float *x, const long len);
+long portWrite(const ubx_port_t *p, const double *x, const long len);
+long portWrite(const ubx_port_t *p, const int32_t *x, const long len);
+long portWrite(const ubx_port_t *p, const int64_t *x, const long len);
+long portWrite(const ubx_port_t *p, const uint32_t *x, const long len);
+long portWrite(const ubx_port_t *p, const uint64_t *x, const long len);
 
 template <typename T>
 ros::Subscriber makeSub(ros::NodeHandle *nh,
